% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chars_collapse}
\alias{chars_collapse}
\title{Collapse character vectors}
\usage{
chars_collapse(x, collapse = ", ")
}
\arguments{
\item{x}{A character vector}

\item{collapse}{A separator to collapse the \code{x} vector.}
}
\value{
A character vector of length=1 collapsing the character vector with \code{collapse} if
the original length is more than 1. Used to create more legible error messages.
}
\description{
A conditional wrapper around \code{\link[base]{paste}}.
}
\keyword{internal}
