% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_digits.R
\name{check_digits}
\alias{check_digits}
\title{Check digits parameter}
\usage{
check_digits(digits)
}
\arguments{
\item{digits}{Digit input to check for validity.}

\item{data_table}{A symmetric input output table, a use table or a supply 
table.}
}
\value{
An error if the digits are not \code{NULL} or an integer value.
}
\description{
This is an internal function to determine if the rounding can go ahead.
}
\examples{
check_digits (1)
}
