% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_industry_to_product.R
\name{convert_industry_to_product}
\alias{convert_industry_to_product}
\title{Convert industry columns to product columns}
\usage{
convert_industry_to_product(data_table)
}
\arguments{
\item{data_table}{A table where the names of the columns are denoted
with the \code{ind_ava} or \code{ind_use} or \code{ind_use} vocabularies, for
example, agriculture is denoted with \code{A01}.}
}
\value{
A converted, named vector or table where the columns are
denoted with \code{prd_ava} codes, i.e., \code{A01} is converted to \code{CPA_A01}.
}
\description{
A utility function to adjust column names for supplementary data,
for example, air pollution data, where the industry codes are denoted
with \code{ind_ava} or \code{ind_use} (or similar codes).
}
\details{
For example, \code{A01} (agriculture) is converted to \code{CPA_A01} (product
of agriculture).

This corresponds to the bridge matrix (B) defined in Eurostat
Manual §15.4.2,  used to transform industry-based extensions to
product-based form.
}
\examples{
data_table <- data.frame(
  ind_use = "A01",
  A01 = 1, A02 = 3, TOTAL = 4,
  HH = 2
)
convert_industry_to_product(data_table)
}
\seealso{
\code{\link[=supplementary_add]{supplementary_add()}}
}
