% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment_get.R
\name{employment_get}
\alias{employment_get}
\title{Get employment data}
\source{
Eurostat dataset:
\url{https://ec.europa.eu/eurostat/web/products-datasets/-/lfsq_egan22d}
}
\usage{
employment_get(
  geo,
  year = "2010",
  sex = "Total",
  age = "Y_GE15",
  labelling = "iotables",
  data_directory = NULL,
  force_download = FALSE
)
}
\arguments{
\item{geo}{A two-letter country code (Eurostat style).}

\item{year}{Year of employment data (>= 2008, when NACE Rev. 2 was
introduced).}

\item{sex}{Employment by sex. Defaults to \code{"Total"}. Use \code{"Females"}/\code{"F"} or
\code{"Males"}/\code{"M"}.}

\item{age}{Eurostat age code. Defaults to \code{"Y_GE15"}. Any Eurostat code can
be supplied.}

\item{labelling}{Either \code{"iotables"}, \code{"prod_na"} (product × product) or
\code{"induse"} (industry × industry).}

\item{data_directory}{Optional path to save or load pre-processed employment
data.}

\item{force_download}{Logical. If \code{TRUE}, forces a fresh download even if a
file exists locally.}
}
\value{
A \code{data.frame} formatted to conform with Eurostat SIOTs, with
auxiliary metadata.
}
\description{
Download Eurostat employment data for a country and arrange it to match 64x64
symmetric input–output tables (SIOTs).
}
\details{
Currently only works with product × product tables.
}
\examples{
\dontrun{
employment <- employment_get(
  geo = "CZ",
  year = "2010",
  sex = "Total",
  age = "Y_GE15",
  data_directory = NULL,
  force_download = TRUE
)
}
}
\seealso{
Other import functions: 
\code{\link{airpol_get}()},
\code{\link{iotables_download}()},
\code{\link{iotables_metadata_get}()},
\code{\link{iotables_read_tempdir}()}
}
\concept{import functions}
