% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-croatia_employment_2013.R
\docType{data}
\name{employment_hr}
\alias{employment_hr}
\title{Croatian employment data for the year 2013}
\format{A data frame with 107 observations in 2 variables:
\describe{
  \item{code}{Short labels}
  \item{iotables_label}{iotables sytle labels}
  \item{employment}{Employment in the sector in Croatia, not in thousands!}
}}
\usage{
data(employment_hr)
}
\description{
Aggregate Croatian detailed employment statistics into the Croatian (EU standard)
Symmetric input-output table format.
}
\keyword{Croatia,}
\keyword{data,}
\keyword{datasets,}
\keyword{employmnet}
