% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eurostat_filtered.R
\name{get_eurostat_data}
\alias{get_eurostat_data}
\title{Use this function for non-SIOT data like air pollutants or
employment}
\usage{
get_eurostat_data(id, filters = list(), type = "code", lang = "en", ...)
}
\description{
Use this function for non-SIOT data like air pollutants or
employment
}
\keyword{internal}
