% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_saved_table.R
\name{get_saved_table}
\alias{get_saved_table}
\alias{find_saved_table}
\alias{get_package_iots}
\title{Get an input-output table from a downloaded bulk file}
\usage{
get_saved_table(labelled_io_data, geo, year, unit, stk_flow)

find_saved_table(labelled_io_data, geo, unit, year, stk_flow)

get_package_iots(source_input)
}
\description{
Get an input-output table from a downloaded bulk file
}
\keyword{internal}
