% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_column_find.R
\name{household_column_find}
\alias{household_column_find}
\title{Return the position of final household expenditure}
\usage{
household_column_find(data_table)
}
\arguments{
\item{data_table}{A symmetric input output table, a use table or a supply 
table.}
}
\value{
An integer value with the final household expenditure. Returns 
\code{NULL} if not found.
}
\description{
Return the position of final household expenditure
}
\examples{
household_column_find( iotable_get ( source = 'germany_1990') )
}
