% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_column_get.R
\name{household_column_get}
\alias{household_column_get}
\title{Return final household expenditure}
\usage{
household_column_get(data_table)
}
\arguments{
\item{data_table}{A symmetric input output table, a use table or a supply 
table.}
}
\value{
The column containing final househodl expenditure. If not found
\code{NULL} is returned.
}
\description{
Return final household expenditure
}
\examples{
household_column_get( iotable_get ( source = 'germany_1990') )
}
