% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotables-package.R
\docType{package}
\name{iotables-package}
\alias{iotables}
\alias{iotables-package}
\title{iotables: Reproducible Input-Output Economics Analysis, Economic and Environmental Impact Assessment with Empirical Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Pre-processing and basic analytical tasks related to working with Eurostat's symmetric input-output tables and provide basic input-output economics calculations. The package is part of rOpenGov \url{http://ropengov.github.io/} to open source open government initiatives.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iotables.dataobservatory.eu/}
  \item Report bugs at \url{https://github.com/rOpenGov/iotables/issues}
}

}
\author{
\strong{Maintainer}: Daniel Antal \email{daniel.antal@dataobservatory.eu} (\href{https://orcid.org/0000-0001-7513-6760}{ORCID})

Other contributors:
\itemize{
  \item Kasia Kulma (\href{https://orcid.org/0000-0002-2952-9720}{ORCID}) [contributor]
  \item Pyry Kantanen (\href{https://orcid.org/0000-0003-2853-2765}{ORCID}) [contributor]
}

}
\keyword{internal}
