% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotables.R
\docType{package}
\name{iotables}
\alias{iotables}
\title{iotables: Working Programatically With Symmetric Input-Output Tables}
\description{
Pre-processing and basic analytical tasks related to working with 
Eurostat's symmetric input-output tables and provide basic 
input-output economics calculations. The package is 
a part of rOpenGov <http://ropengov.github.io/> for 
open source open government initiatives.
}
\section{Metadata datasets}{

Data files that contain descriptive metadata for a correct reproduction
of the symmetric input-output tables.  The analytical functions use 
matrix equations that require a precise column and row order for each
table.
}

\section{Validation datasets}{

Data files that replicate published input-output tables with analysis.
These files are used to validate the correct working of the analytical
functions.
}

\section{Croatia data files}{

These are Croatia's symmetric input-output tables for the year 2010, when 
the country was not yet an EU member state.
}

