% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontieff_matrix_create.R
\name{leontieff_matrix_create}
\alias{leontieff_matrix_create}
\title{Create a Leontieff matrix}
\usage{
leontieff_matrix_create(technology_coefficients_matrix)
}
\arguments{
\item{technology_coefficients_matrix}{A technology coefficient matrix created by the \code{\link{input_coefficient_matrix_create}}.}
}
\description{
Create a Leontieff matrix from technology matrix after some basic error handling.
}
\examples{
de_use <- use_table_get ( source = "germany_1990", geo = "DE",
               year = 1990, unit = "MIO_EUR", 
               households = FALSE, labelling = "iotables")

de_output <- output_get ( source = "germany_1990", geo = "DE",
               year = 1990, unit = "MIO_EUR",
               households = FALSE, labelling = "iotables")

de_coeff <- input_coefficient_matrix_create( de_use, de_output, digits = 4)

L <- iotables::leontieff_matrix_create( technology_coefficients_matrix = de_coeff )
}
