% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-metadata.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata}
\format{A data frame with 10 variables. #' 
\describe{
  \item{variable}{Eurostat vocabulary source, i.e. t_rows, t_cols, prod_na, induse}
  \item{group}{ Informal labelling for macroeconomic groups }
  \item{code}{Eurostat labels}
  \item{label}{Eurostat label descriptions}
  \item{quadrant}{Where to place the data from a long-form raw data file}  
  \item{account_group}{Different from Eurostat tables, in thousand national currency units.}
  \item{digit_1}{ third digit for ordering}
  \item{digit_2}{ fourth digit for ordering}
  \item{numeric_label}{ ordering from quadrant, account_group, digit_1, digit_2}     
  \item{iotables_label}{Custom, machine_readable snake format variable names}                     
}}
\source{
\href{https://www.dzs.hr/Hrv_Eng/publication/2015/12-01-04_01_2015.xlsx}{Eurostat Manual of Supply, Use and Input-Output Tables} Updated 17 December 2015.
}
\usage{
data(metadata)
}
\description{
An arrangement of the Eurostat national accounts vocabulary, used to correctly
order wide format rows and columns from bulk long-form tables.
}
\keyword{data,}
\keyword{datasets,}
\keyword{input-output}
\keyword{metadata,}
\keyword{table,}
\keyword{vocabulary}
