% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg_enumerate}
\alias{msg_enumerate}
\title{Enumerate character values}
\usage{
msg_enumerate(x, sep = ", ")
}
\arguments{
\item{x}{A character vector}

\item{collapse}{A separator to collapse the \code{x} vector.}
}
\value{
A character vector of length=1 collapsing the character vector with \code{collapse} if
the original length is more than 1.
}
\description{
A conditional wrapper around \code{basee::paste}.
}
\keyword{internal}
