% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplier_create.R
\name{multiplier_create}
\alias{multiplier_create}
\title{Solve a basic equation}
\usage{
multiplier_create(input_vector = NULL, Im = NULL, digits = NULL)
}
\arguments{
\item{input_vector}{An input matrix or vector created by the input_indicator_create () function.}

\item{Im}{The Leontieff inverse as a named object created by the leontieff_inverse_create () function.}

\item{digits}{Rounding digits, if omitted, no rounding takes place. #' @examples 
data (germany_1990)
get_input_flow(df = germany_1990, geo = 'DE', year = 1990,
               unit = "M_EUR", named = FALSE)}
}
\description{
This function is in fact a wrapper around the equation_solve () function, adding a name to the multiplier.
}
