% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_netherlands_2006.R
\docType{data}
\name{netherlands_2006}
\alias{netherlands_2006}
\title{Simple input-output table for the Netherlands, 2006.}
\format{A data frame with 14 observations and 13 variables.}
\source{
Source: Input-Output Multipliers Specification Sheet and Supporting
Material in the Spicosa Project Report
}
\usage{
data(netherlands_2006)
}
\description{
This simplified SIOT is taken from the  Science Policy Integration for Coastal
Systems Assessment project's input-output multiplier specification sheet.
It is used as a simple example SIOT for controlled analytical results.
The column names were slightly altered to resemble more the current Eurostat
conventions and the main example dataset \code{\link{germany_1990}}.
}
\keyword{Netherlands}
\keyword{data,}
\keyword{datasets,}
\keyword{input-output}
\keyword{table,}
