% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_coefficient_matrix_create.R
\name{output_coefficient_matrix_create}
\alias{output_coefficient_matrix_create}
\title{Create an input coefficient matrix}
\usage{
output_coefficient_matrix_create(input_flow, digits = NULL)
}
\arguments{
\item{input_flow}{An input flow matrix created with the \code{\link{use_table_get}} function.}

\item{digits}{An integer showing the precision of the technology matrix in 
digits. Default is \code{NULL} when no rounding is applied.}
}
\description{
Create an output coefficient matrix from the input flow matrix. If there
are zero values in the output vector, they will be changed to 
0.000001 and you will get a warning. Some analytical equations cannot be 
solved with zero elements. You either have faulty input data, or you have 
to use some sort of data modification to carry on your analysis.
}
\details{
An alternative that is not implemented here, because it requires analytical
judgment, is the aggregation of elements into larger ones that are no longer
equal to zero, i.e. merging an industry or product class that has a positive 
value with another industry or product class that is zero.
}
\examples{
de_use <- use_table_get ( source = "germany_1990", geo = "DE",
                          year = 1990, unit = "MIO_EUR", 
                          households = FALSE, labelling = "iotables")
                          
output_coefficient_matrix_create ( input_flow = de_use )
}
