% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_get.R
\name{output_get}
\alias{output_get}
\title{Get an output vector}
\usage{
output_get(data_table)
}
\arguments{
\item{data_table}{A symmetric input-output table or use table retrieved 
by the  
\code{\link{iotable_get}} function.}
}
\value{
A data frame with the vector of multipliers and the an 
auxiliary metadata column (for joining with other matrixes.)
}
\description{
This is a wrapper function around the \code{\link{primary_input_get}} 
function.
}
\examples{
output_get ( data_table = iotable_get () )
}
\seealso{
Other iotables processing functions: 
\code{\link{conforming_vector_create}()},
\code{\link{household_column_get}()},
\code{\link{primary_input_get}()}
}
\concept{iotables processing functions}
