% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-primary_inputs.R
\docType{data}
\name{primary_inputs}
\alias{primary_inputs}
\title{Primary input abbreviations}
\format{
A data frame with 105 rows (including empty ones) and 2 variables.
\describe{
  \item{t_rows2}{Eurostat code of the input.}
  \item{t_rows2_lab}{Labelling of the input by Eurostat.}
  \item{source}{Eurostat / DZS }
  \item{indicator}{Human readable abbreviation}
}
}
\usage{
data("croatia_employment_aggregation")
}
\description{
Only currently used primary inputs. Abbreviations for filtering.
}
\keyword{Eurostat,}
\keyword{data,}
\keyword{datasets,}
\keyword{input-output}
\keyword{inputs}
\keyword{metadata,}
\keyword{table,}
