% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrant_separator_find.R
\name{quadrant_separator_find}
\alias{quadrant_separator_find}
\title{Determine the end of Quadrant I and III.}
\usage{
quadrant_separator_find(data_table, include_total = FALSE)
}
\arguments{
\item{data_table}{A symmetric input output table, a use table or a supply
table.}

\item{include_total}{Should the total (intermediary) output column be
included \code{TRUE} or excluded (\code{FALSE}, default)?}
}
\value{
An integer value with the last column of Quadrant I and III. If
the last column is not found, \code{2} is returned with a warning to avoid
stopping a pipeline.
}
\description{
This is an internal function to determine where to separate quadrants if
necessary.
}
\keyword{internal}
