% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk2010_get.R
\name{uk_2010_get}
\alias{uk_2010_get}
\title{Get United Kingdom Input-Output Analytical Tables, 2010}
\source{
\href{https://www.ons.gov.uk/file?uri=/economy/nationalaccounts/supplyandusetables/datasets/ukinputoutputanalyticaltablesdetailed/2010detailed/ukioanalyticaltablesio1062010detailedpubversion.xls}
{ukioanalyticaltablesio1062010detailedpubversion.xls}
}
\usage{
uk_2010_get(path = NULL)
}
\arguments{
\item{path}{A path to the downloaded file, if already exists, given with
\code{\link{base::file.path}} function.}
}
\description{
This function will retrieve any primary input from the input-output table.
United Kingdom Input-Output Analytical Tables, 2010												
(consistent with UK National Accounts Blue Book 2013 &
 UK Balance of Payments Pink Book 2013)							
by Richard Wild.
}
\examples{
\dontrun{
uk2010 <- uk_2010_get()
}
}
