% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_2010_results_get.R
\name{uk_2010_results_get}
\alias{uk_2010_results_get}
\title{Get United Kingdom Multipliers and Effects, 2010}
\source{
\href{https://www.ons.gov.uk/file?uri=/economy/nationalaccounts/supplyandusetables/datasets/ukinputoutputanalyticaltablesdetailed/2010detailed/ukioanalyticaltablesio1062010detailedpubversion.xls}{ukioanalyticaltablesio1062010detailedpubversion.xls}
}
\usage{
uk_2010_results_get(path = NULL)
}
\arguments{
\item{path}{A path to the downloaded file, if already exists, given with
\code{file.path()} function.}
}
\description{
This function will retrieve the published effects and multipliers from the
United Kingdom Input-Output Analytical Tables, 2010
(consistent with UK National Accounts Blue Book 2013 &
UK Balance of Payments Pink Book 2013)
by Richard Wild.
}
\examples{
\dontrun{
uk_results <- iotables:::uk_2010_results_get()
}
}
