% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_source}
\alias{validate_source}
\title{Validate source Parameter}
\usage{
validate_source(source)
}
\arguments{
\item{source}{Possible data sources.}
}
\description{
Validate source Parameter
}
\keyword{internal}
