% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_transpose.R
\name{vector_transpose}
\alias{vector_transpose}
\title{Transpose a vector}
\usage{
vector_transpose(
  data_table,
  names_to = "nace_r2",
  values_to = "value",
  .keep = FALSE
)
}
\arguments{
\item{data_table}{A matrix or vector that should have a key column.}

\item{names_to}{Defaults to \code{'nace_r2'}.}

\item{values_to}{Defaults to \code{'value'}.}
}
\value{
A logical variable of length 1, \code{TRUE} or \code{FALSE}.
}
\description{
Many vectors (indicators, multipliers) are create in the wide form to confom matrixes in 
analytical functions.  For printing it is more useful to have them in long form.
}
\examples{
vector_transpose ( data.frame (indicator = "my_inidcator", 
                               agriculture = 0.0123,
                               manufacturing = 0.1436,
                               trade = 0.0921
                                ))
}
