% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_linkages.R
\name{backward_linkages}
\alias{backward_linkages}
\title{Backward linkages}
\usage{
backward_linkages(Im)
}
\arguments{
\item{Im}{A Leontief inverse matrix created by the
\code{\link{leontief_inverse_create}} function.}
}
\value{
The vector of industry (product) backward linkages in a wide
data.frame class, following the column names of the Leontief 
inverse matrix.
}
\description{
Indicate the interconnection of a particular sector to other sectors
from which it purchases inputs (demand side). When a sector increases its output, 
it will increase the total (intermediate) demand on all other sectors, which is measured
by backward linkages.
}
\details{
Backward linkages are defined as the column sum of the Leontief inverse, in line with 
the Eurostat Manual of Supply, Use and
Input-Output Tables (see p506-507.) and the 
Handbook on Supply and Use Tables and Input-Output Tables with Extensions and Applications of 
the United Nations (see p636,)
}
\examples{
de_coeff <- input_coefficient_matrix_create( iotable_get(), 
                                             digits = 4 )
I <- leontief_inverse_create (de_coeff)
backward_linkages (I)
}
\seealso{
Other linkage functions: 
\code{\link{forward_linkages}()}
}
\concept{linkage functions}
