% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chars_collapse}
\alias{chars_collapse}
\title{Collapse character vectors}
\usage{
chars_collapse(x, collapse = ", ")
}
\arguments{
\item{x}{A character vector.}

\item{collapse}{A separator string used if \code{x} has more than one element.}
}
\value{
A character string of length one if \code{x} has length > 1, otherwise the
original vector unchanged.
}
\description{
A wrapper around \code{\link[base:paste]{base::paste()}} that conditionally collapses a
character vector. If the vector length is greater than one, it is collapsed
using the supplied separator. Used internally to create legible error
messages.
}
\keyword{internal}
