% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_matrix_create.R
\name{coefficient_matrix_create}
\alias{coefficient_matrix_create}
\title{Create a coefficient matrix}
\usage{
coefficient_matrix_create(siot, total = "output", digits = NULL,
  empy_row_remove = TRUE)
}
\arguments{
\item{siot}{A symmetric input-output table retrieved by the
\code{\link{iotable_get}} function.}

\item{total}{Usually an output vector with a key column, defaults to 
\code{"output"} which equals \code{"P1"} or \code{"output_bp"}.
You can use other rows for comparison, for example \code{"TS_BP"} 
if it exists in the matrix.}

\item{digits}{An integer showing the precision of the technology matrix in 
digits. Default is \code{NULL} when no rounding is applied.}

\item{empy_row_remove}{Defaults to \code{TRUE}, if \code{FALSE}, empty
primary input rows are returned. Empty row/column pairs from the
technology matrix are always removed to keep the table symmetric and to 
avoid division by zero in analytical functions.}
}
\value{
A data.frame that contains the matrix of  \code{siot} devided by \code{total}
with a key column.
}
\description{
Create a coefficient matrix from a Symmetric Input-Output Table. The
coefficient matrix is related by default to output, but you can change this
to total supply or other total aggregate if it exists in your table.
}
\details{
The empty columns are always removed to avoid division by zero errors. The empty
rows are removed by default, but you can optionally keep them.
}
\examples{
coefficient_matrix_create ( siot = iotable_get ( source = "germany_1990"), 
                            total = "output", 
                            digits = 4 )
}
\references{
See \href{https://webarchive.nationalarchives.gov.uk/20160114044923/http://www.ons.gov.uk/ons/rel/input-output/input-output-analytical-tables/2010/index.html}{United Kingdom Input-Output Analytical Tables 2010}
for explanation on the use of the Coefficient matrix.
}
