% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-croatia_employment_2013.R
\docType{data}
\name{croatia_employment_2013}
\alias{croatia_employment_2013}
\title{Croatian employment data for the year 2013}
\format{
A data frame with 107 observations in 2 variables:
\describe{
  \item{code}{Short labels}
  \item{iotables_row}{iotables style labels}
  \item{employment}{Employment in the sector in Croatia, not in thousands!}
}
}
\usage{
data(croatia_employment_2013)
}
\description{
Aggregate Croatian detailed employment statistics into the Croatian (EU standard)
Symmetric input-output table format.
}
\seealso{
Other Croatia 2010 datasets: 
\code{\link{croatia_2010_1700}},
\code{\link{croatia_2010_1800}},
\code{\link{croatia_2010_1900}},
\code{\link{croatia_employment_aggregation}},
\code{\link{primary_inputs}}
}
\concept{Croatia 2010 datasets}
\keyword{datasets}
