% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_effects_create.R
\name{direct_effects_create}
\alias{direct_effects_create}
\title{Create direct effects}
\usage{
direct_effects_create(input_requirements, inverse, digits = NULL)
}
\arguments{
\item{input_requirements}{A matrix or vector created by
\code{\link{input_indicator_create}}}

\item{inverse}{A Leontief-inverse created by
\code{\link{leontief_inverse_create}}.}

\item{digits}{Rounding digits, defaults to \code{NULL}, in which case
no rounding takes place.}
}
\value{
A data.frame containing the direct effects and the necessary
metadata to sort them or join them with other matrixes.
}
\description{
The function creates the effects.
}
\examples{
nl <- netherlands_2000

input_coeff_nl <- input_coefficient_matrix_create(
  data_table = netherlands_2000,
  households = FALSE
)

compensation_indicator <- input_indicator_create(netherlands_2000, "compensation_employees")

I_nl <- leontief_inverse_create(input_coeff_nl)

direct_effects_create(
  input_requirements = compensation_indicator,
  inverse = I_nl
)
}
\seealso{
Other indicator functions: 
\code{\link{coefficient_matrix_create}()},
\code{\link{input_indicator_create}()}
}
\concept{indicator functions}
