% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-employment_metadata.R
\docType{data}
\name{employment_metadata}
\alias{employment_metadata}
\title{Employment metadata}
\format{
A data frame with 6 variables.  
\describe{
  \item{emp_code}{code used in the employment statistics}
  \item{code}{Eurostat labels for SIOTs corresponding to emp_code}
  \item{label}{Eurostat label descriptions for SIOTs corresponding to emp_code}
  \item{variable}{Eurostat vocabulary source, i.e. t_rows, t_cols, prod_na, induse}
  \item{group}{Different from Eurostat tables, in thousand national currency units.}
  \item{iotables_label}{Custom, machine_readable snake format variable names}                     
}
}
\usage{
data(employment_metadata)
}
\description{
An arrangement of the Eurostat national accounts vocabulary to match with 
employment statistics data.
}
\seealso{
Other Metadata datasets: 
\code{\link{metadata_uk_2010}},
\code{\link{metadata}}
}
\concept{Metadata datasets}
\keyword{datasets}
