% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_germany.R
\docType{data}
\name{germany_1995}
\alias{germany_1995}
\title{Simple input–output table for Germany, 1995}
\format{
A data frame with 228 observations and 11 variables:
\describe{
\item{prod_na}{Technology codes in row names, following Eurostat
conventions.}
\item{prod_na_lab}{Longer labels corresponding to the \code{t_rows2} codes.}
\item{induse}{Column labels, following Eurostat conventions with
minor differences.}
\item{iotables_row}{Row labels (used in key columns) in \code{iotables}
abbreviations.}
\item{iotables_col}{Column labels in \code{iotables} abbreviations.}
\item{values}{Table values in millions of euros.}
\item{unit}{Unit identifier. Always \code{"MIO_EUR"}.}
\item{unit_lab}{Label for the unit: "million euros". Eurostat usually
provides both euro and national currency values.}
\item{geo}{Country code for Germany (\code{"DE"}).}
\item{geo_lab}{Country name (\code{"Germany"}).}
\item{time}{Reference year of the SIOT.}
}
}
\source{
Eurostat (2008). \emph{Eurostat Manual of Supply, Use and Input–Output
Tables}, p. 482. Available from:
\url{https://ec.europa.eu/eurostat/documents/3859598/5902113/KS-RA-07-013-EN.PDF}
}
\usage{
germany_1995
}
\description{
Replication dataset from the Eurostat Manual, Table 15.4:
\emph{Input–output table of domestic output at basic prices (Version A)}.
}
\details{
This dataset is used for testing and documentation purposes. It is a
well-documented example taken from the Eurostat Manual. The table has
been reformatted into the same long-form tidy structure as the Eurostat
database SIOTs.

The labels were slightly adjusted to reflect the transition from the
ESA95 vocabulary to ESA2010 since the publication of the Manual.
}
\seealso{
Other validation datasets: 
\code{\link{germany_airpol}},
\code{\link{uk_test_results}}
}
\concept{validation datasets}
\keyword{datasets}
