% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metadata.R
\name{getdata}
\alias{getdata}
\title{Various internal functions to work with IOT metadata, including the labelling 
vocabularies, row and column ordering. 
None of these functions should be exported.}
\usage{
getdata(...)
}
\description{
Various internal functions to work with IOT metadata, including the labelling 
vocabularies, row and column ordering. 
None of these functions should be exported.
}
\keyword{internal}
