% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghosh_inverse_create.R
\name{ghosh_inverse_create}
\alias{ghosh_inverse_create}
\title{Create the inverse of a Ghosh-matrix.

Create the Ghosh-inverse from the output coefficients.}
\usage{
ghosh_inverse_create(output_coefficients_matrix, digits = NULL)
}
\arguments{
\item{output_coefficients_matrix}{A technology coefficient matrix created
by the \code{\link{output_coefficient_matrix_create}}.}

\item{digits}{An integer showing the precision of the technology matrix in 
digits. Default is \code{NULL} when no rounding is applied.}
}
\description{
Create the inverse of a Ghosh-matrix.

Create the Ghosh-inverse from the output coefficients.
}
\examples{
om <- output_coefficient_matrix_create( 
  data_table = iotable_get(), 
  households = FALSE)
ghosh_inverse_create( output_coefficients_matrix = om )
}
\seealso{
Other analytic object functions: 
\code{\link{input_flow_get}()},
\code{\link{leontieff_inverse_create}()},
\code{\link{leontieff_matrix_create}()}
}
\concept{analytic object functions}
