% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_indicator_create.R
\name{input_indicator_create}
\alias{input_indicator_create}
\title{Create input indicator(s)}
\usage{
input_indicator_create(data_table, input_vector = c("gva_bp",
  "net_tax_production"), digits = NULL, households = FALSE,
  indicator_names = NULL)
}
\arguments{
\item{digits}{Rounding digits, if omitted, no rounding takes place.}

\item{indicator_names}{The names of new indicators. Defaults to \code{NULL} when 
the names in the key colum of \code{input_matrix} will be used to create the 
indicator names.}

\item{input_matrix}{A named (primary) input(s) vector or matrix created by \code{\link{primary_input_get}}}

\item{output_vector}{A named output vector created by \code{\link{output_get}}.}
}
\value{
A tibble (data frame) containing  \code{input_matrix} devided by the \code{output_vector}
with a key column for products or industries.
}
\description{
The function creates the input indicators from the inputs and the outputs.
}
\examples{
 
input_indicator_create( data_table = iotable_get(), 
                        input_vector = "compensation_employees",
                        digits = 4)
}
