% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_multipliers_create.R
\name{input_multipliers_create}
\alias{input_multipliers_create}
\title{Create input multipliers}
\usage{
input_multipliers_create(direct_effects, inverse, digits = NULL,
  multiplier_names = NULL)
}
\arguments{
\item{direct_effects}{A direct effects matrix created by 
\code{\link{direct_supply_effects_create}}}

\item{inverse}{A Leontieff-inverse created by \code{\link{leontieff_inverse_create}}.}

\item{digits}{Rounding digits, defaults to \code{NULL}, in which case 
no rounding takes place.}
}
\description{
The function creates the input multipliers.  See Eurostat Manual p500-501..
}
\examples{
 
direct_effects_de <- coefficient_matrix_create(data_table = data_table, 
                                              total = 'output', 
                                              return = 'primary_inputs') 

#Remove the total column from direct effects to conform with the inverse:
direct_effects_de <- dplyr::select(direct_effects_de, -total)

multipliers <- input_multipliers_create(
 direct_effects = direct_effects_de,
 inverse = I_de)
}
