% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R, R/iotables-package.R
\docType{package}
\name{iotables-package}
\alias{iotables-package}
\title{Package globals}
\description{
Symbols declared here avoid R CMD check "no visible binding for global variable"
notes. This file is transitional: as functions are refactored to use
\code{.data$var} or \code{.env$var} pronouns, entries can be removed.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Pre-processing and basic analytical tasks for working with Eurostat's symmetric input–output tables, and basic input–output economics calculations. Part of rOpenGov \url{https://ropengov.github.io/} for open source open government initiatives.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iotables.dataobservatory.eu/}
  \item \url{https://github.com/rOpenGov/iotables}
  \item Report bugs at \url{https://github.com/rOpenGov/iotables/issues}
}


Useful links:
\itemize{
  \item \url{https://iotables.dataobservatory.eu/}
  \item \url{https://github.com/rOpenGov/iotables}
  \item Report bugs at \url{https://github.com/rOpenGov/iotables/issues}
}

}
\author{
\strong{Maintainer}: Daniel Antal \email{daniel.antal@dataobservatory.eu} (\href{https://orcid.org/0000-0001-7513-6760}{ORCID})

Other contributors:
\itemize{
  \item Kasia Kulma (\href{https://orcid.org/0000-0002-2952-9720}{ORCID}) [contributor]
  \item Pyry Kantanen (\href{https://orcid.org/0000-0003-2853-2765}{ORCID}) [contributor]
}

}
\keyword{internal}
