% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotables.R
\docType{package}
\name{iotables}
\alias{iotables}
\title{iotables: Working with Symmetric Input-Output Tables Programatically}
\description{
Pre-processing and basic analytical tasks related to working with 
Eurostat's symmetric input-output tables and provide basic 
input-output economics calculations. The package is 
a part of rOpenGov <http://ropengov.github.io/> for 
open source open government initiatives.
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Croatia data files}{

These are Croatia's symmetric input-output tables for the year 2010, when 
the country was not yet an EU member state.
}

\section{Example data files}{

Data files that replicate published input-output tables with analysis.
These files are used to validate the correct working of the analytical
functions.
}

