% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotables.R
\docType{package}
\name{iotables}
\alias{iotables}
\title{iotables: Reproducible Input–Output Analysis for Eurostat and National Data}
\description{
Tools for downloading, structuring, and analysing input–output and
supply–use tables, with metadata aligned to Eurostat’s ESA 2010 framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iotables.dataobservatory.eu/}
  \item \url{https://github.com/rOpenGov/iotables}
  \item Report bugs at \url{https://github.com/rOpenGov/iotables/issues}
}

}
\author{
\strong{Maintainer}: Daniel Antal \email{daniel.antal@dataobservatory.eu} (\href{https://orcid.org/0000-0001-7513-6760}{ORCID})

Other contributors:
\itemize{
  \item Pyry Kantanen (\href{https://orcid.org/0000-0003-2853-2765}{ORCID}) [contributor]
}

}
