% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_key_column_present}
\alias{is_key_column_present}
\title{Is a keyword present in a key column?}
\usage{
is_key_column_present(data_table, potential_keywords = NULL)
}
\arguments{
\item{data_table}{A matrix or vector that should have a key column.}

\item{potential_keywords}{Potential keywords that should be present in the key column, defaults to
\code{NULL} in which case it will asserted that the first column is not numeric.}
}
\value{
A logical variable of length 1, \code{TRUE} or \code{FALSE}.
}
\description{
Is a keyword present in a key column?
}
\keyword{internal}
