% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontieff_matrix_create.R
\name{leontieff_matrix_create}
\alias{leontieff_matrix_create}
\title{Create a Leontieff matrix}
\usage{
leontieff_matrix_create(technology_coefficients_matrix)
}
\arguments{
\item{technology_coefficients_matrix}{A technology coefficient matrix created
by the \code{\link{input_coefficient_matrix_create}} or 
\code{\link{output_coefficient_matrix_create}}.}
}
\description{
Create a Leontieff matrix from technology matrix after some basic error 
handling. Most likely you will need this function as a step to invoke the 
function to create its inverse: \code{\link{leontieff_inverse_create}}.
}
\examples{
tm <- input_flow_get ( 
  data_table = iotable_get(), 
  households = FALSE)
L <- iotables::leontieff_matrix_create( technology_coefficients_matrix = tm )
}
