% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_round.R
\name{matrix_round}
\alias{matrix_round}
\title{Round all matrix values to required number of digits.}
\usage{
matrix_round(data_table, digits = 0)
}
\arguments{
\item{data_table}{A symmetric input output table, a use table or a supply 
table.}

\item{digits}{An integer number, defaults to \code{0}.}
}
\value{
The matrix, with the intact key column and the numeric columns rounded.
}
\description{
Round all matrix values to required number of digits.
}
\examples{
matrix_round( iotable_get ( source = 'germany_1990') )
}
