% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplier_create.R
\name{multiplier_create}
\alias{multiplier_create}
\title{Create multipliers}
\usage{
multiplier_create(input_vector = NULL, Im = NULL,
  multiplier_name = "multiplier", digits = NULL)
}
\arguments{
\item{input_vector}{An input matrix or vector created by the input_indicator_create () function.}

\item{Im}{The Leontieff inverse as a named object created by the leontieff_inverse_create () function.}

\item{multiplier_name}{A variable name to be given to the returned multipliers. Defaults to 'multiplier'.}

\item{digits}{Rounding digits, if omitted, no rounding takes place.}
}
\description{
This function is in fact a wrapper around the equation_solve () function, adding a name to the multiplier.
}
\examples{
de_emp <- primary_input_get ( input = "employment_total",
           source = "germany_1990", geo = "DE",
           year = 1990,  
           households = FALSE, labelling = "iotables")

de_emp_indicator <- input_indicator_create (de_emp, de_output)

employment_multipliers <- multiplier_create ( 
           input_vector    = de_emp_indicator,
           Im              = I_de,
           multiplier_name =  "employment_multiplier", 
           digits = 4 )

}
