% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_zero_columns_find.R
\name{non_zero_columns_find}
\alias{non_zero_columns_find}
\title{Find Non-zero Columns}
\usage{
non_zero_columns_find(data_table)
}
\arguments{
\item{data_table}{A symmetric input output table, a use table or a supply 
table.}
}
\value{
A vector of \code{TRUE} and \code{FALSE} values for the table.
}
\description{
This is an internal function to help finding empty columns and rows in 
symmetric tables.
}
\keyword{internal}
