% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_coefficients_create.R
\name{output_coefficients_create}
\alias{output_coefficients_create}
\title{Create output coefficients}
\usage{
output_coefficients_create(data_table, output)
}
\arguments{
\item{data_table}{A matrix or vector that should have a key column.}

\item{output}{An output vector created with \code{output_get}.}
}
\description{
Create coefficients from supplementary data.
}
\examples{
emissions_de <- germany_airpol[, -3] 
emissions_de <- vector_transpose_wider( names_from = "iotables_col", values_from = "value")
output_bp <- output_get (iotable_get())
output_coefficients_create (emissions_de, output_bp)
}
