% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_get.R
\name{output_get}
\alias{output_get}
\title{Get an output vector}
\usage{
output_get(data_table = NULL, labelling = "iotables")
}
\arguments{
\item{data_table}{A symmetric input-output table or use table retrieved by the
\code{\link{iotable_get}} function.}
}
\description{
This is a wrapper function around the \code{\link{primary_input_get}} 
function.
}
\examples{
output_get ( data_table = iotable_get () )
}
