% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_input_get.R
\name{primary_input_get}
\alias{primary_input_get}
\title{Get primary inputs}
\usage{
primary_input_get(siot, primary_input = "compensation_employees")
}
\arguments{
\item{siot}{A symmetric input-output table created by 
\code{\link{iotable_get}}.}
}
\description{
This function will retrieve any primary input from the input-output table.
}
\examples{
comp_employees_de <- primary_input_get(
                            siot = iotable_get( "germany_1990"), 
                            primary_input = "compensation_employees")
                            )
}
