% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_input_get.R
\name{primary_input_get}
\alias{primary_input_get}
\title{Get primary inputs}
\usage{
primary_input_get(data_table, primary_input = "compensation_employees")
}
\arguments{
\item{data_table}{A symmetric input-output table, a use table, or a 
supply table retrieved by the  \code{\link{iotable_get}} function.}

\item{primary_input}{The primary input to be returned from the table.}
}
\description{
This function will retrieve any primary input from the input-output table.
}
\examples{
comp_employees_de <- primary_input_get(
  data_table = iotable_get(), 
  primary_input = "compensation_employees")                             )
}
