% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_table.R
\name{round_table}
\alias{round_table}
\title{Systematically round values in a table.}
\usage{
round_table(data_table, digits = NULL)
}
\arguments{
\item{data_table}{A symmetric input-output table, a use table, a supply table,
a margin or tax table.}

\item{digits}{}
}
\value{
A tibble/data.frame with a key row and a symmetric matix, 
after removing all empty columns and rows at the same time.
}
\description{
This is an internal function to do the rounding on all numeric elements of
the table.
}
\examples{
round_table( iotable_get ( source = "germany_1990") )
}
