% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdir_data.R
\name{tempdir_data}
\alias{tempdir_data}
\title{Download to and retrieve from the temporary directory a Eurostat dataset}
\usage{
tempdir_data(id)
}
\arguments{
\item{id}{The id of a Eurostat product.}
}
\description{
Download to and retrieve from the temporary directory a Eurostat dataset
}
\keyword{internal}
