% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-uk_test_results.R,
%   R/data-uk_test_results_2010.R
\docType{data}
\name{uk_test_results}
\alias{uk_test_results}
\title{Multipliers and effects (product) for testing
from the United Kingdom Input-Output Analytical Tables, 2010}
\format{A data frame with 10 variables.
\describe{
  \item{uk_row}{The UK row identifier. Dots and '&' converted to '-'.} 
  \item{indicator}{The name of the indicator, i.e. Excel sheet.}
  \item{Output multiplier}{The imported Output multipliers}
  \item{Employment cost multiplier}{The imported Employment cost multipliers.}
  \item{GVA multiplier}{The imported GVA multipliers.}
  \item{GVA effects}{The imported GVA effects.}
}}
\usage{
data(uk_2010_data)

data(uk_2010_data)
}
\description{
The Excel-imported UK data.

The Excel-imported UK data.
}
\keyword{UK}
\keyword{data,}
\keyword{datasets,}
\keyword{effects,}
\keyword{input-output}
\keyword{multipliers,}
\keyword{table,}
