% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-uk_test_results.R
\docType{data}
\name{uk_test_results}
\alias{uk_test_results}
\title{Multipliers and effects (product) for testing
from the United Kingdom Input-Output Analytical Tables, 2010}
\format{
A data frame with 12 variables.
\describe{
  \item{uk_row_label}{The UK row label} 
  \item{Output multiplier}{The imported Output multipliers}
  \item{output_multiplier_rank}{The imported ranking of output multipliers}
  \item{Employment cost multiplier}{The imported Employment cost multipliers.}
  \item{employment_cost_multiplier}{The imported ranking of Employment cost multipliers.}
  \item{Employment cost effects}{The imported Employment cost multipliers.}
  \item{employment_cost_effects_rank}{The imported ranking of employment cost multipliers.}
  \item{GVA effects}{The imported GVA effects.}
  \item{gva_effects_rank}{The imported ranking GVA effects.}
  \item{gva_multiplier_rank}{The imported ranking GVA multipliers.}
  \item{GVA multiplier}{The imported GVA multipliers.}
  \item{indicator}{Indicator names.}
}
}
\usage{
data(uk_test_results)
}
\description{
The Excel-imported UK data.
}
\seealso{
Other Validation datasets: 
\code{\link{germany_1995}},
\code{\link{germany_airpol}},
\code{\link{netherlands_2006}},
\code{\link{uk_2010_data}}
}
\concept{Validation datasets}
\keyword{datasets}
