% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_table_get.R
\name{use_table_get}
\alias{use_table_get}
\title{Create a use (input flow) matrix}
\usage{
use_table_get(source = "germany_1990", geo = "DE", year = 1990,
  unit = "MIO_EUR", households = FALSE, labelling = "iotables")
}
\arguments{
\item{source}{A data source, for example "naio_10_cp1700".}

\item{geo}{A country code or a country name.  For example, "SK" or as "Slovakia".}

\item{year}{A numeric variable containing the year. Defaults to 2010, because this year has the most data.}

\item{unit}{A character string containing the currency unit, defaults to "MIO_NAC" (million national currency unit). The alternative is "MIO_EUR".}

\item{households}{If you need to make household demand endogenous, or "close the households off", TRUE selects 
wages and final household consumption. This is needed for induced-effects calculations.}

\item{labelling}{Defaults to "iotables" which gives standard row and column names regardless of the
source of the table, or if it is a product x product, industry x industry or product x industry table.
The alternative is "short" which is the original short row or column code of Eurostat or OECD.}
}
\description{
Create a use (input flow) matrix
}
\examples{
use_de <- use_table_get( source = "germany_1990", geo = 'DE', year = 1990,
                       unit = "MIO_EUR", labelling  = 'iotables')
\dontrun{
sk_use <- use_table_get ( source = "naio_cp17_r2", geo = "SK",
                          year = 2010, unit = "MIO_EUR", households = "FALSE",
                          labelling = "iotables" )
 }
}
